+++ CHANGELOG.md
@@ -24,6 +24,7 @@ parameter
 arbitrarily applied mapper pass.
 - The measure instruction accepts an axis parameter
 - `MeasureDecomposer` to decompose arbitrary measurements to a decomposition of single-qubit gates and a +Z measurement.
+- `Can2CZDecomposer` to decompose arbitrary two-qubit gates. 
 
 ## [ 0.9.0 ] - [ 2025-12-19 ]